move some functions
